cambios en la forma de obtener la distancia

+++ src/main/java/org/circle8/entity/Punto.java
@@ -8,6 +8,6 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @Builder
 public class Punto {
-	public double latitud;
-	public double longitud;
+	public float latitud;
+	public float longitud;
 }
